swap: move shared test helpers to separate file
