+++ swap/swap_test.go
@@ -23,7 +23,6 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"encoding/json"
-	"errors"
 	"flag"
 	"fmt"
 	"io/ioutil"
@@ -37,18 +36,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ethersphere/swarm/network"
-
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind/backends"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/log"
-	"github.com/ethereum/go-ethereum/p2p"
 	"github.com/ethereum/go-ethereum/p2p/enode"
-	"github.com/ethereum/go-ethereum/p2p/simulations/adapters"
 	"github.com/ethereum/go-ethereum/rpc"
 	contractFactory "github.com/ethersphere/go-sw3/contracts-v0-2-0/simpleswapfactory"
 	cswap "github.com/ethersphere/swarm/contracts/swap"
@@ -75,21 +68,6 @@ type booking struct {
 	peer   *protocols.Peer
 }
 
-// swapTestBackend encapsulates the SimulatedBackend and can offer
-// additional properties for the tests
-type swapTestBackend struct {
-	*backends.SimulatedBackend
-	factoryAddress common.Address // address of the SimpleSwapFactory in the simulated network
-	tokenAddress   common.Address // address of the token in the simulated network
-	// the async cashing go routine needs synchronization for tests
-	cashDone chan struct{}
-}
-
-func (b *swapTestBackend) Close() error {
-	// Do not close SimulatedBackend as it is a global instance.
-	return nil
-}
-
 func TestMain(m *testing.M) {
 	exitCode := m.Run()
 	// Close the global default backend
@@ -102,41 +80,6 @@ func init() {
 	testutil.Init()
 	mrand.Seed(time.Now().UnixNano())
 	swapLog = log.Root()
-}
-
-var defaultBackend = backends.NewSimulatedBackend(core.GenesisAlloc{
-	ownerAddress:       {Balance: big.NewInt(1000000000000000000)},
-	beneficiaryAddress: {Balance: big.NewInt(1000000000000000000)},
-}, 8000000)
-
-// newTestBackend creates a new test backend instance
-func newTestBackend(t *testing.T) *swapTestBackend {
-	t.Helper()
-	// commit the initial "pre-mined" accounts (issuer and beneficiary addresses)
-	defaultBackend.Commit()
-
-	// deploy the ERC20-contract
-	// ignore receipt because if there is no error, we can assume everything is fine on a simulated backend
-	tokenAddress, _, _, err := contractFactory.DeployERC20Mintable(bind.NewKeyedTransactor(ownerKey), defaultBackend)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defaultBackend.Commit()
-
-	// deploy a SimpleSwapFactoy
-	// ignore receipt because if there is no error, we can assume everything is fine on a simulated backend
-	// ignore factory instance, because the address is all we need at this point
-	factoryAddress, _, _, err := contractFactory.DeploySimpleSwapFactory(bind.NewKeyedTransactor(ownerKey), defaultBackend, tokenAddress)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defaultBackend.Commit()
-
-	return &swapTestBackend{
-		SimulatedBackend: defaultBackend,
-		factoryAddress:   factoryAddress,
-		tokenAddress:     tokenAddress,
-	}
 }
 
 type storeKeysTestCase struct {
@@ -848,135 +791,6 @@ func TestRestoreBalanceFromStateStore(t *testing.T) {
 	}
 }
 
-// During tests, because the cashing in of cheques is async, we should wait for the function to be returned
-// Otherwise if we call `handleEmitChequeMsg` manually, it will return before the TX has been committed to the `SimulatedBackend`,
-// causing subsequent TX to possibly fail due to nonce mismatch
-func testCashCheque(s *Swap, cheque *Cheque) {
-	cashCheque(s, cheque)
-	// send to the channel, signals to clients that this function actually finished
-	if stb, ok := s.backend.(*swapTestBackend); ok {
-		if stb.cashDone != nil {
-			stb.cashDone <- struct{}{}
-		}
-	}
-}
-
-// newDefaultParams creates a set of default params for tests
-func newDefaultParams(t *testing.T) *Params {
-	t.Helper()
-	baseKey := make([]byte, 32)
-	_, err := rand.Read(baseKey)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return &Params{
-		BaseAddrs:           network.NewBzzAddr(baseKey, nil),
-		LogPath:             "",
-		PaymentThreshold:    int64(DefaultPaymentThreshold),
-		DisconnectThreshold: int64(DefaultDisconnectThreshold),
-	}
-}
-
-// newBaseTestSwapWithParams creates a swap with the given params
-func newBaseTestSwapWithParams(t *testing.T, key *ecdsa.PrivateKey, params *Params, backend *swapTestBackend) (*Swap, string) {
-	t.Helper()
-	dir, err := ioutil.TempDir("", "swap_test_store")
-	if err != nil {
-		t.Fatal(err)
-	}
-	stateStore, err := state.NewDBStore(dir)
-	if err != nil {
-		t.Fatal(err)
-	}
-	log.Debug("creating simulated backend")
-	owner := createOwner(key)
-	swapLog = newSwapLogger(params.LogPath, params.BaseAddrs)
-	factory, err := cswap.FactoryAt(backend.factoryAddress, backend)
-	if err != nil {
-		t.Fatal(err)
-	}
-	swap := newSwapInstance(stateStore, owner, backend, 10, params, factory)
-	return swap, dir
-}
-
-// create a test swap account with a backend
-// creates a stateStore for persistence and a Swap account
-func newBaseTestSwap(t *testing.T, key *ecdsa.PrivateKey, backend *swapTestBackend) (*Swap, string) {
-	params := newDefaultParams(t)
-	return newBaseTestSwapWithParams(t, key, params, backend)
-}
-
-// create a test swap account with a backend
-// creates a stateStore for persistence and a Swap account
-// returns a cleanup function
-func newTestSwap(t *testing.T, key *ecdsa.PrivateKey, backend *swapTestBackend) (*Swap, func()) {
-	t.Helper()
-	usedBackend := backend
-	if backend == nil {
-		usedBackend = newTestBackend(t)
-	}
-	swap, dir := newBaseTestSwap(t, key, usedBackend)
-	clean := func() {
-		swap.Close()
-		// only close if created by newTestSwap to avoid double close
-		if backend != nil {
-			backend.Close()
-		}
-		os.RemoveAll(dir)
-	}
-	return swap, clean
-}
-
-type dummyPeer struct {
-	*protocols.Peer
-}
-
-// creates a dummy protocols.Peer with dummy MsgReadWriter
-func newDummyPeer() *dummyPeer {
-	return newDummyPeerWithSpec(nil)
-}
-
-// creates a dummy protocols.Peer with dummy MsgReadWriter
-func newDummyPeerWithSpec(spec *protocols.Spec) *dummyPeer {
-	id := adapters.RandomNodeConfig().ID
-	rw := &dummyMsgRW{}
-	protoPeer := protocols.NewPeer(p2p.NewPeer(id, "testPeer", nil), rw, spec)
-	dummy := &dummyPeer{
-		Peer: protoPeer,
-	}
-	return dummy
-}
-
-// creates cheque structure for testing
-func newTestCheque() *Cheque {
-	cheque := &Cheque{
-		ChequeParams: ChequeParams{
-			Contract:         testChequeContract,
-			CumulativePayout: uint64(42),
-			Beneficiary:      beneficiaryAddress,
-		},
-		Honey: uint64(42),
-	}
-
-	return cheque
-}
-
-// creates a randomized cheque structure for testing
-func newRandomTestCheque() *Cheque {
-	amount := mrand.Intn(100)
-
-	cheque := &Cheque{
-		ChequeParams: ChequeParams{
-			Contract:         testChequeContract,
-			CumulativePayout: uint64(amount),
-			Beneficiary:      beneficiaryAddress,
-		},
-		Honey: uint64(amount),
-	}
-
-	return cheque
-}
-
 // tests if encodeForSignature encodes the cheque as expected
 func TestChequeEncodeForSignature(t *testing.T) {
 	expectedCheque := newTestCheque()
@@ -1101,24 +915,6 @@ func TestVerifyContractNotDeployedByFactory(t *testing.T) {
 	}
 }
 
-// setupContractTest is a helper function for setting up the
-// blockchain wait function for testing
-func setupContractTest() func() {
-	// we overwrite the waitForTx function with one which the simulated backend
-	// immediately commits
-	currentWaitFunc := cswap.WaitFunc
-	// we also need to store the previous cashCheque function in case this is called multiple times
-	currentCashCheque := defaultCashCheque
-	defaultCashCheque = testCashCheque
-	// overwrite only for the duration of the test, so...
-	cswap.WaitFunc = testWaitForTx
-	return func() {
-		// ...we need to set it back to original when done
-		cswap.WaitFunc = currentWaitFunc
-		defaultCashCheque = currentCashCheque
-	}
-}
-
 // TestFactoryAddressForNetwork tests that an address is found for ropsten
 // and no address if for network 32145
 func TestFactoryAddressForNetwork(t *testing.T) {
@@ -1148,96 +944,6 @@ func TestFactoryVerifySelf(t *testing.T) {
 	if factory.VerifySelf() != nil {
 		t.Fatal(err)
 	}
-}
-
-// when testing, we don't need to wait for a transaction to be mined
-func testWaitForTx(ctx context.Context, backend cswap.Backend, tx *types.Transaction) (*types.Receipt, error) {
-
-	var stb *swapTestBackend
-	var ok bool
-	if stb, ok = backend.(*swapTestBackend); !ok {
-		return nil, errors.New("not the expected test backend")
-	}
-	stb.Commit()
-
-	receipt, err := backend.TransactionReceipt(ctx, tx.Hash())
-	if err != nil {
-		return nil, err
-	}
-	if receipt.Status != types.ReceiptStatusSuccessful {
-		return nil, cswap.ErrTransactionReverted
-	}
-	return receipt, nil
-}
-
-// deploy for testing (needs simulated backend commit)
-func testDeployWithPrivateKey(ctx context.Context, backend cswap.Backend, privateKey *ecdsa.PrivateKey, ownerAddress common.Address, depositAmount *big.Int) (cswap.Contract, error) {
-	opts := bind.NewKeyedTransactor(privateKey)
-	opts.Context = ctx
-
-	var stb *swapTestBackend
-	var ok bool
-	if stb, ok = backend.(*swapTestBackend); !ok {
-		return nil, errors.New("not the expected test backend")
-	}
-
-	factory, err := cswap.FactoryAt(stb.factoryAddress, stb)
-	if err != nil {
-		return nil, err
-	}
-
-	// setup the wait for mined transaction function for testing
-	cleanup := setupContractTest()
-	defer cleanup()
-	contract, err := factory.DeploySimpleSwap(opts, ownerAddress, big.NewInt(int64(defaultHarddepositTimeoutDuration)))
-	if err != nil {
-		return nil, err
-	}
-	stb.Commit()
-
-	// send money into the new chequebook
-	token, err := contractFactory.NewERC20Mintable(stb.tokenAddress, stb)
-	if err != nil {
-		return nil, err
-	}
-
-	tx, err := token.Mint(bind.NewKeyedTransactor(ownerKey), contract.ContractParams().ContractAddress, depositAmount)
-	if err != nil {
-		return nil, err
-	}
-	stb.Commit()
-	receipt, err := stb.TransactionReceipt(ctx, tx.Hash())
-	if err != nil {
-		return nil, err
-	}
-
-	if receipt.Status != 1 {
-		return nil, errors.New("token transfer reverted")
-	}
-
-	return contract, nil
-}
-
-// deploy for testing (needs simulated backend commit)
-func testDeploy(ctx context.Context, swap *Swap, depositAmount *big.Int) (err error) {
-	swap.contract, err = testDeployWithPrivateKey(ctx, swap.backend, swap.owner.privateKey, swap.owner.address, depositAmount)
-	return err
-}
-
-// newTestSwapAndPeer is a helper function to create a swap and a peer instance that fit together
-// the owner of this swap is the beneficiaryAddress
-// hence the owner of this swap would sign cheques with beneficiaryKey and receive cheques from ownerKey (or another party) which is NOT the owner of this swap
-func newTestSwapAndPeer(t *testing.T, key *ecdsa.PrivateKey) (*Swap, *Peer, func()) {
-	swap, clean := newTestSwap(t, key, nil)
-	// owner address is the beneficiary (counterparty) for the peer
-	// that's because we expect cheques we receive to be signed by the address we would issue cheques to
-	peer, err := swap.addPeer(newDummyPeer().Peer, ownerAddress, testChequeContract)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// we need to adjust the owner address on swap because we will issue cheques to beneficiaryAddress
-	swap.owner.address = beneficiaryAddress
-	return swap, peer, clean
 }
 
 // TestPeerSetAndGetLastReceivedCheque tests if a saved last received cheque can be loaded again later using the peer functions
@@ -1618,18 +1324,4 @@ func TestAvailableBalance(t *testing.T) {
 		t.Fatalf("availableBalance not equal to deposited minus withdraw. availableBalance: %d, deposit minus withdrawn: %d", availableBalance, depositAmount.Uint64()-withdrawAmount.Uint64())
 	}
 
-}
-
-// dummyMsgRW implements MessageReader and MessageWriter
-// but doesn't do anything. Useful for dummy message sends
-type dummyMsgRW struct{}
-
-// ReadMsg is from the MessageReader interface
-func (d *dummyMsgRW) ReadMsg() (p2p.Msg, error) {
-	return p2p.Msg{}, nil
-}
-
-// WriteMsg is from the MessageWriter interface
-func (d *dummyMsgRW) WriteMsg(msg p2p.Msg) error {
-	return nil
 }
